requirements.txt
Django==2.0.4
django-tastypie==0.14.1
python-dateutil==2.7.2
gunicorn==19.6.0